FamilyHarmonyAdvisor/API/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
spacy==3.7.2
transformers==4.35.0
torch==2.1.0